requirements/requirements.txt
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
feedparser==5.2.1
gTTS==2.0.4
gTTS-token==1.1.3
idna==2.8
numpy==1.16.2
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3
croniter==0.3.30
pyzmq==18.1.0